+++ docs/conf.py
@@ -1,16 +1,14 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-import os
-
 extensions = [
     "sphinx.ext.autodoc",
     "sphinx.ext.autosummary",
+    "sphinx.ext.napoleon",
     "sphinx.ext.coverage",
     "sphinx.ext.doctest",
     "sphinx.ext.extlinks",
     "sphinx.ext.ifconfig",
-    "sphinx.ext.napoleon",
     "sphinx.ext.todo",
     "sphinx.ext.viewcode",
 ]
@@ -20,7 +18,7 @@ project = "TimeThese"
 year = "2020"
 author = "Joachim W. Bargsten"
 copyright = "{0}, {1}".format(year, author)
-version = release = "0.0.4"
+version = release = "0.0.5"
 
 pygments_style = "trac"
 templates_path = ["."]
@@ -28,11 +26,6 @@ extlinks = {
     "issue": ("https://github.com/jwbargsten/python-timethese/issues/%s", "#"),
     "pr": ("https://github.com/jwbargsten/python-timethese/pull/%s", "PR #"),
 }
-# on_rtd is whether we are on readthedocs.org
-on_rtd = os.environ.get("READTHEDOCS", None) == "True"
-
-if not on_rtd:  # only set the theme if we're building docs locally
-    html_theme = "sphinx_rtd_theme"
 
 html_use_smartypants = True
 html_last_updated_fmt = "%b %d, %Y"